fix if

+++ entrypoint.sh
@@ -96,8 +96,5 @@ terraform ${TF_VERB} ${TF_PLAN} ${TF_VARSFILE} ${TF_AUTOAPPROVE} ${TF_OUT}
 # Copy tfplan to github workspace
 if [ "$TF_VERB" = "plan" ]
 then
-    if [ ! -f /github/workspace/$INPUT_PLANFILE ]
-    then
-      cp tfplan /github/workspace/$INPUT_PLANFILE
-    fi
+    [ ! -f /github/workspace/$INPUT_PLANFILE ] && cp tfplan /github/workspace/$INPUT_PLANFILE
 fi
